Fixed prefix stripping issues. Cf https://github.com/APSL/puput/pull/82#issuecomment-266992663
