+++ puput/utils.py
@@ -2,6 +2,7 @@
 
 from six import string_types
 from importlib import import_module
+from django.core.urlresolvers import reverse
 
 
 def import_model(path_or_callable):
@@ -16,3 +17,18 @@ def import_model(path_or_callable):
 def get_image_model_path():
     from django.conf import settings
     return getattr(settings, 'WAGTAILIMAGES_IMAGE_MODEL', 'wagtailimages.Image')
+
+
+def strip_prefix_and_ending_slash(path):
+    """
+    If puput and wagtail are registered with a prefix, it needs to be removed
+    so the 'entry_page_serve_slug' or 'blog_page_feed_slug' resolutions can work.
+    Ex, here with a dynamic (i18n_patterns()) + a static prefix :
+    urlpatterns += i18n_patterns(
+        url(r'^blah/', include('puput.urls')),
+        url(r'^blah/', include(wagtail_urls)),
+    )
+    The prefix is simply the root where Wagtail page are served.
+    https://github.com/torchbox/wagtail/blob/stable/1.8.x/wagtail/wagtailcore/urls.py#L36
+    """
+    return path.replace(reverse('wagtail_serve', args=[""]), '', 1).rstrip("/")
